Add files via upload
Cooling degree days: A cooling degree day (CDD) is a measurement designed to quantify the demand for energy needed to cool a building. It is the number of degrees that a day's average temperature is higher than a defined temperature, which is the temperature above which buildings need to be cooled.